Graded HW5
Docstrings are present for functions, but are not detailed enough. There are many numbers and variables that are unclear and should be explained through in-line comments. Keep comments professional, as Github could serve as part of your portfoilo. **Documentation: +0.8**

Make sure all code is enveloped in a function of some sort for good practice. Repeating chunks of code should be placed in loops when possible. However, code is clear and easy to follow. **Style: +1**

Code runs as intended. Game is really hard to play ): ... But fun! **Functionality: +3**

Nicely done! :)
Grade: 4.8/5